requirements.txt
gym==0.26.2
numpy==1.23.1
pygame==2.1.2
tensorflow-gpu==2.10.0
tflearn==0.5.0
tqdm==4.64.1